+++ src/fs/create_dir.rs
@@ -16,8 +16,8 @@ use crate::{
 
 /// CreateDir traits for flipper filesystem
 pub trait FsCreateDir {
-    /// Creates a directory at a path.
-    fn fs_create_dir(&mut self, path: impl AsRef<Path>) -> Result<()>;
+    /// Creates a directory at a path. Returns weather the path existed. False = did not exist; True = Already existed.
+    fn fs_create_dir(&mut self, path: impl AsRef<Path>) -> Result<bool>;
 }
 
 impl<T> FsCreateDir for T
@@ -25,16 +25,16 @@ where
     T: TransportRaw<proto::Main, proto::Main, Err = Error> + CommandIndex + std::fmt::Debug,
 {
     #[doc(alias = "fs_mkdir")]
-    fn fs_create_dir(&mut self, path: impl AsRef<Path>) -> Result<()> {
+    fn fs_create_dir(&mut self, path: impl AsRef<Path>) -> Result<bool> {
         let path = os_str_to_str(path.as_ref().as_os_str())?.to_string();
 
         debug!("creating directory at {path}");
 
         match self.send_and_receive(Request::StorageMkdir(path)) {
-            Ok(_) => Ok(()),
+            Ok(_) => Ok(false),
             Err(Error::Rpc(crate::rpc::error::Error::StorageError(
                 crate::rpc::error::StorageError::AlreadyExists,
-            ))) => Ok(()),
+            ))) => Ok(true),
 
             Err(e) => Err(e),
         }