+++ src/fs/tar.rs
@@ -1,4 +1,4 @@
-//! FsTarExtract module. Flipper-side .tgz extraction.
+//! FsTarExtract module. Flipper-side .tar extraction.
 
 use std::path::Path;
 
@@ -16,15 +16,15 @@ use crate::{
 
 /// MD5 trait
 pub trait FsTarExtract {
-    /// Extracts a .tgz on the flipper from path -> out
-    fn fs_extract_tgz(&mut self, path: impl AsRef<Path>, out: impl AsRef<Path>) -> Result<()>;
+    /// Extracts a .tar, NOTE: NOT A TGZ, on the flipper from path -> out
+    fn fs_extract_tar(&mut self, path: impl AsRef<Path>, out: impl AsRef<Path>) -> Result<()>;
 }
 
 impl<T> FsTarExtract for T
 where
     T: TransportRaw<proto::Main, proto::Main, Err = Error> + CommandIndex + std::fmt::Debug,
 {
-    fn fs_extract_tgz(&mut self, path: impl AsRef<Path>, out: impl AsRef<Path>) -> Result<()> {
+    fn fs_extract_tar(&mut self, path: impl AsRef<Path>, out: impl AsRef<Path>) -> Result<()> {
         let path = os_str_to_str(path.as_ref().as_os_str())?.to_string();
         let out = os_str_to_str(out.as_ref().as_os_str())?.to_string();